+++ packages/react/src/components/billing-page.tsx
@@ -13,7 +13,7 @@ import {
 } from '../FlowgladContext'
 import { useCallback } from 'react'
 import { CreateCheckoutSessionParams } from '@flowglad/shared'
-import { FlowgladTheme } from '@/FlowgladTheme'
+import { FlowgladTheme } from '../FlowgladTheme'
 
 const SectionTitle = ({
   children,
@@ -113,7 +113,7 @@ export function BillingPage({
   if (!billing.loadBilling || !billing.loaded || !billing.catalog) {
     return <div>Loading...</div>
   }
-
+  console.log('darkMode', darkMode)
   return (
     <FlowgladTheme darkMode={darkMode}>
       <div