+++ packages/server/src/FlowgladServer.ts
@@ -211,6 +211,7 @@ export class FlowgladServer {
   public createAddPaymentMethodCheckoutSession = async (
     params: CreateAddPaymentMethodCheckoutSessionParams
   ): Promise<FlowgladNode.CheckoutSessions.CheckoutSessionCreateResponse> => {
+    this.flowgladNode.subscriptions.list()
     return this.createCheckoutSession({
       ...params,
       type: 'add_payment_method',