+++ packages/express/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @flowglad/express
 
+## 0.10.2
+
+### Patch Changes
+
+- Fix file path for FlowgladTheme import
+- Updated dependencies
+  - @flowglad/server@0.10.2
+  - @flowglad/shared@0.10.2
+  - @flowglad/react@0.10.2
+
+## 0.10.1
+
+### Patch Changes
+
+- Move FlowgladTheme to billing-page only for now
+- Updated dependencies
+  - @flowglad/react@0.10.1
+  - @flowglad/server@0.10.1
+  - @flowglad/shared@0.10.1
+
 ## 0.10.0
 
 ### Minor Changes